* Fix progress when installing/removing software (Bug #32148)
git-svn-id: svn+ssh://billy.knut.univention.de/var/univention/svn/dev/branches/ucs-3.2/ucs-3.2-0@43807 1b283449-9f2f-0410-b571-8eb228e1a901

+++ base/univention-system-setup/umc/python/setup/util.py
@@ -306,7 +306,7 @@ class ProgressParser( object ):
 		match = ProgressParser.STEP.match( line )
 		if match is not None:
 			try:
-				self.current.step = int( match.groups()[ 0 ] )
+				self.current.step = float( match.groups()[ 0 ] )
 				if self.current.step > self.current.steps:
 					self.current.step = self.current.steps
 				return True
